Refactor and harden PLC communication logic
This commit addresses a comprehensive code review by implementing several key improvements:

- **Security**: The critical security vulnerability of disabling TLS certificate validation (`rejectUnauthorized: false`) has been addressed. The application now attempts to load a user-provided `plc-cert.pem` for secure HTTPS communication. A detailed explanation has been added to the README.

- **Resilience**: The polling mechanism is now more resilient to transient network errors. It will attempt to reconnect several times before declaring the connection lost, preventing disconnects from minor network blips.

- **Modularity**: The monolithic `main.js` has been refactored into smaller, single-responsibility modules:
    - `src/api/siemens-plc.js` for the PLC API class.
    - `src/utils/file-parsers.js` for file parsing functions.
    - `src/main-process/ipc-handlers.js` for all IPC handlers.

- **Data Validation**: Data type validation for write operations has been moved from the frontend to the backend (`ipc-handlers.js`). This centralizes the logic and ensures that only correctly-formatted data is sent to the PLC.

- **Documentation**: The README.md has been significantly updated to include details on features, usage, and the new security configuration.
